+++ src/main/java/com/sparta/cream/dto/response/YourPaymentListResponse.java
@@ -1,21 +1,32 @@
 package com.sparta.cream.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.sparta.cream.domain.entity.Payment;
 
 import lombok.Getter;
 
+/**
+ * 사용자 본인의 결제 목록 응답 데이터를 나타내는 DTO 클래스입니다.
+ * <p>
+ * id, merchantUid, productName, amount, paidAt, status를 포함합니다.
+ * </p>
+ *
+ * @author 변채주
+ * @version 1.0
+ * @since 2026. 02. 12.
+ */
 @Getter
 public class YourPaymentListResponse {
 	private final Long id;
 	private final String merchantUid;
 	private final String productName;
-	private final Long amount;
+	private final BigDecimal amount;
 	private final LocalDateTime paidAt;
 	private final String status;
 
-	public YourPaymentListResponse(Long id, String merchantUid, String productName, Long amount, LocalDateTime paidAt,
+	public YourPaymentListResponse(Long id, String merchantUid, String productName, BigDecimal amount, LocalDateTime paidAt,
 		String status) {
 		this.id = id;
 		this.merchantUid = merchantUid;
